refactor: extract meeting completion boundaries
Keep meeting edge handling and repair-safe receipt emission in focused owners, and document the remaining exhaustive-switch/API-boundary line-count deltas.\n\nVerification: 123 focused backend tests; 37 focused Swift tests.
