+++ .github/scripts/product_file_line_count_ratchet_baseline/backend-routers.json
@@ -2,7 +2,7 @@
   "files": {
     "backend/routers/apps.py": 2425,
     "backend/routers/chat.py": 1714,
-    "backend/routers/developer.py": 2064,
+    "backend/routers/developer.py": 2071,
     "backend/routers/mcp_sse.py": 1895,
     "backend/routers/sync.py": 2083,
     "backend/routers/users.py": 2174
@@ -10,7 +10,7 @@
   "raise_justifications": {
     "backend/routers/apps.py": "POST /v1/apps/enable reads is_setup_completed through _setup_completed_from_response so a developer-controlled body that is non-JSON or not an object returns the intended 400 instead of a 500 (46 prod occurrences Jul 21-28); the guard stays in the route that owns the enable response contract. +5 keeps deleted-source rejection at the anonymous owner-migration admission boundary so a fresh UID cannot claim state from a fenced owner (SCA-280). +28 gives /v2/apps/search the name-OR-description predicate its `q` parameter documents plus the client-parity relevance tier, next to the one handler that owns them rather than importing search policy into utils/apps.py (#11292).",
     "backend/routers/chat.py": "The HTTP and streaming PTT routes own the raw-audio and final-transcript boundaries; dev-only allowlisted capture therefore starts and persists at those two route seams while serialization, bounds, redaction, and export stay in testing.parity_pack_v0.live_capture. +33 emits done: from a staged typed answer when present and keeps the canned fallback when error: arrives without an answer so persona failures still finalize (SCA-296). +5 classifies staged typed stream-error answers as journey failure + shared fallback while keeping the typed done: copy (SCA-296). +35 keeps an in-memory typed done: frame when Firestore persistence fails so text clients are not left with only an earlier error: frame (SCA-296). +9 isolates post-persist app-usage analytics failures so they cannot mint a second client-visible message id (SCA-296).",
-    "backend/routers/developer.py": "Developer memory create and batch routes must record their accepted write result at the authoritative API boundary; the shared dev-only allowlisted serializer/exporter lives in testing.parity_pack_v0.live_capture. +40 keeps category-filtered compatibility pagination at the route's released response boundary while all storage access remains in MemoryService.",
+    "backend/routers/developer.py": "Developer memory create and batch routes must record their accepted write result at the authoritative API boundary; the shared dev-only allowlisted serializer/exporter lives in testing.parity_pack_v0.live_capture. +40 keeps category-filtered compatibility pagination at the route's released response boundary while all storage access remains in MemoryService. +7 carries the desktop conversation-role marker through the existing from-segments request and invokes the extracted meeting-arrival adapter at each idempotent completion boundary.",
     "backend/routers/mcp_sse.py": "The MCP SSE tool handler owns the successful memory-write response boundary; its one shared dev-only allowlisted capture call records accepted memory without duplicating serialization or export logic. +24 fences legacy-key, OAuth, and authorization-code issuance at the route's credential admission seams while offloading request validation, Firebase, and Firestore work to their owned executors (SCA-280). +42 keeps hosted MCP tool-call analytics at the authoritative execution boundary while the shared helper owns privacy filtering, PostHog delivery, and connector-ready operation mapping (SCA-294). +23 records every execution outcome and preserves authorization semantics independently of JSON-RPC error codes (SCA-294). +24 applies the account-cutover fence to authenticated MCP principals at their existing authorization boundary. +21 merges transcript-chunk hits into search_conversations and returns match_snippets so MCP search covers transcript evidence (#6621).",
     "backend/routers/sync.py": "Fresh Sync's daily ceiling remains beside the existing fresh-admission gates, before staged audio or durable worker work is created. Lifecycle-fenced workers are terminalized at this Cloud Tasks boundary so released WAL clients do not retry an owner they no longer own; +2 for the queued-dispatch-lost stale finalization path (#10033). +30 applies account-generation admission and background cutover fencing at the sync mutation and worker boundaries.",
     "backend/routers/users.py": "Account-deletion outcome telemetry at the users router boundary for PostHog ownership repair; repaired non-empty webhook URLs must also clear the persisted disabled state and failure health at the owning save boundary (#7519). +56 hosts the return-only AI user profile synthesis route at the boundary that already owns the ai-profile read/write contract; both prompt stages and the consolidation live in utils/llm/ai_user_profile.py. +5 moves the synthesis route onto the LLM executor and refuses a paywalled account before the provider call. +8 serves the primary-language picker list from GET /v1/users/available-languages so a language can be added without an app release; the response models live in models/users.py to keep the router growth to the route itself. +6 aligns the Task assistant prompt request bound with the shipped desktop default at the existing assistant-settings schema boundary."
