+++ .github/scripts/product_file_line_count_ratchet_baseline/desktop-swift-floatingcontrolbar.json
@@ -1,6 +1,6 @@
 {
   "files": {
-    "desktop/macos/Desktop/Sources/FloatingControlBar/AgentPill.swift": 2375,
+    "desktop/macos/Desktop/Sources/FloatingControlBar/AgentPill.swift": 2376,
     "desktop/macos/Desktop/Sources/FloatingControlBar/FloatingControlBarView.swift": 2704,
     "desktop/macos/Desktop/Sources/FloatingControlBar/FloatingControlBarWindow.swift": 5101,
     "desktop/macos/Desktop/Sources/FloatingControlBar/PushToTalkManager.swift": 2807,
@@ -8,7 +8,7 @@
     "desktop/macos/Desktop/Sources/FloatingControlBar/RealtimeHubSession.swift": 1658
   },
   "raise_justifications": {
-    "desktop/macos/Desktop/Sources/FloatingControlBar/AgentPill.swift": "Agent completion presentation now follows the canonical terminal lifecycle used by PTT and chat.",
+    "desktop/macos/Desktop/Sources/FloatingControlBar/AgentPill.swift": "Agent completion presentation now follows the canonical terminal lifecycle used by PTT and chat. +1 keeps the new main-chat-only conversation link inert in the floating-agent exhaustive projection.",
     "desktop/macos/Desktop/Sources/FloatingControlBar/FloatingControlBarView.swift": "Onboarding bar glow, reach-error handling, and the notch hover/voice surface-state boundary render on the shared bar surface. The live-suggestion notch card, hover shortcut legend, Second Brain onboarding, reach-error, and notch-moment UI remain distinct overlay branches; the centered notch voice morph stays frontmost through waveform transitions, while the reducer-owned speaking projection keeps response playback synchronized. The default notification and focus-suggestion cards each carry their own accent-icon tile, header, and dismiss layout so proactive nudges are legible at a glance.",
     "desktop/macos/Desktop/Sources/FloatingControlBar/FloatingControlBarWindow.swift": "Typed proactive-notification identity must be journaled beside the floating-bar manager state that presents each card so chat continuity preserves the delivery kind without adding a second writer.",
     "desktop/macos/Desktop/Sources/FloatingControlBar/PushToTalkManager.swift": "Strict concurrency annotations, the capture-free owner-boundary setup seam, and off-main-thread CoreAudio input warmup remain beside PTT entry points. Warm mic keep-alive parks finished capture behind a lock-guarded lease so the next turn reuses the IOProc; the lease and park/adopt state stay beside the capture generation counter (BasedHardware/omi#10499, #10442, #9181).",
