+++ .github/scripts/product_file_line_count_ratchet_baseline/desktop-swift-providers.json
@@ -1,10 +1,10 @@
 {
   "files": {
-    "desktop/macos/Desktop/Sources/Providers/ChatProvider.swift": 6919,
+    "desktop/macos/Desktop/Sources/Providers/ChatProvider.swift": 6922,
     "desktop/macos/Desktop/Sources/Providers/ChatToolExecutor.swift": 3242
   },
   "raise_justifications": {
-    "desktop/macos/Desktop/Sources/Providers/ChatProvider.swift": "+44 for recording a failed turn where the reader will find it. A failed turn left its question orphaned: the assistant row stayed empty, the journal projection deletes an empty `.failed` row as a placeholder, and the only account of the failure was the provider-wide `errorMessage` — dismissible, overwritten by the next turn, gone on relaunch. What the reader is told, which endings count as failures, and what the row persists all moved out to `ChatTurnFailureNotice`; the provider keeps only the three things that are its own state — stamping the row before the journal is finalized, giving the composer its text back, and refusing a second concurrent send out loud instead of silently. The refusal cannot move out: `isSending` and the bridge send lock are this type's state and INV-6 gives them one owner. Also adds `revokeActiveTurn(reason:)`, extracted from `stopAgent`, and routes the three transcript resets through it. Those resets blanked `messages` while leaving `isSending` true, `activeTurnOwner` set and the send lock held, so every later send was rejected as \"already sending\", Stop could not clear it, and the abandoned turn's late result still ran the visible-completion path against rows that no longer existed — a turn that silently vanished while telemetry and the success chime reported success. INV-6 requires one revocation authority in this type: moving it out would give an in-flight turn two owners, which is the defect. +20 computes question-response identity from the canonical owner and conversation session so Swift and the agent kernel accept the same live option receipt.",
+    "desktop/macos/Desktop/Sources/Providers/ChatProvider.swift": "+44 for recording a failed turn where the reader will find it. A failed turn left its question orphaned: the assistant row stayed empty, the journal projection deletes an empty `.failed` row as a placeholder, and the only account of the failure was the provider-wide `errorMessage` — dismissible, overwritten by the next turn, gone on relaunch. What the reader is told, which endings count as failures, and what the row persists all moved out to `ChatTurnFailureNotice`; the provider keeps only the three things that are its own state — stamping the row before the journal is finalized, giving the composer its text back, and refusing a second concurrent send out loud instead of silently. The refusal cannot move out: `isSending` and the bridge send lock are this type's state and INV-6 gives them one owner. Also adds `revokeActiveTurn(reason:)`, extracted from `stopAgent`, and routes the three transcript resets through it. Those resets blanked `messages` while leaving `isSending` true, `activeTurnOwner` set and the send lock held, so every later send was rejected as \"already sending\", Stop could not clear it, and the abandoned turn's late result still ran the visible-completion path against rows that no longer existed — a turn that silently vanished while telemetry and the success chime reported success. INV-6 requires one revocation authority in this type: moving it out would give an in-flight turn two owners, which is the defect. +20 computes question-response identity from the canonical owner and conversation session so Swift and the agent kernel accept the same live option receipt. +3 gives conversationLink the same stable block identity and copy-text projection as the other typed links.",
     "desktop/macos/Desktop/Sources/Providers/ChatToolExecutor.swift": "+12 for save_knowledge_graph discovery_text → backend /v1/knowledge-graph/extract (Luna SSOT). Mapping lives in KnowledgeGraphToolSupport; executor only switches success/failure beside the existing local persist boundary. +14 keeps the save_knowledge_graph compatibility path usable when discovery_text is blank or backend extraction fails, at the executor that owns the tool's input contract."
   },
   "threshold": 1500
