+++ backend/routers/developer.py
@@ -67,7 +67,7 @@ from utils.memory.product_authorization import (
     authorize_memory_external_default_memory_read,
     authorize_memory_external_default_memory_write,
 )
-from utils.task_intelligence.proactive_engine import persist_capture_arrival_intent
+from utils.task_intelligence.proactive_engine import persist_desktop_meeting_arrival
 import logging
 
 logger = logging.getLogger(__name__)
@@ -1463,31 +1463,6 @@ def _conversation_response_from_data(conversation: dict) -> ConversationResponse
     )
 
 
-def _persist_desktop_meeting_arrival(uid: str, request: CreateConversationFromTranscriptRequest, conversation) -> None:
-    external_data = conversation.get('external_data') if isinstance(conversation, dict) else conversation.external_data
-    stored_role = (external_data or {}).get('conversation_role')
-    if stored_role != 'meeting':
-        return
-    source = conversation.get('source') if isinstance(conversation, dict) else conversation.source
-    status = conversation.get('status') if isinstance(conversation, dict) else conversation.status
-    discarded = conversation.get('discarded', False) if isinstance(conversation, dict) else conversation.discarded
-    source = source.value if hasattr(source, 'value') else source
-    status = status.value if hasattr(status, 'value') else status
-    if source != ConversationSource.desktop.value or discarded or status != ConversationStatus.completed.value:
-        return
-    conversation_id = conversation['id'] if isinstance(conversation, dict) else conversation.id
-    structured = conversation.get('structured') if isinstance(conversation, dict) else conversation.structured
-    structured = structured or {}
-    title = structured.get('title') if isinstance(structured, dict) else structured.title
-    overview = structured.get('overview') if isinstance(structured, dict) else structured.overview
-    persist_capture_arrival_intent(
-        uid,
-        conversation_id=conversation_id,
-        summary=title or overview or '',
-        is_desktop_meeting=True,
-    )
-
-
 def _create_conversation_from_segments(
     uid: str,
     request: CreateConversationFromTranscriptRequest,
@@ -1577,7 +1552,7 @@ def _create_conversation_from_segments(
                     request.client_session_id,
                     conversation_id,
                 )
-                _persist_desktop_meeting_arrival(uid, request, existing_conversation)
+                persist_desktop_meeting_arrival(uid, existing_conversation)
                 return _conversation_response_from_data(existing_conversation)
 
     resolved_client_device_id = client_device_id or request.client_device_id
@@ -1615,7 +1590,7 @@ def _create_conversation_from_segments(
                     request.client_session_id,
                     conversation_id,
                 )
-                _persist_desktop_meeting_arrival(uid, request, existing_conversation)
+                persist_desktop_meeting_arrival(uid, existing_conversation)
                 return _conversation_response_from_data(existing_conversation)
             raise HTTPException(status_code=409, detail="Conversation creation already in progress")
     else:
@@ -1653,7 +1628,7 @@ def _create_conversation_from_segments(
         )
         lifecycle_service.persist_processed_conversation(uid, conversation.model_dump())
 
-    _persist_desktop_meeting_arrival(uid, request, conversation)
+    persist_desktop_meeting_arrival(uid, conversation)
 
     return ConversationResponse(
         id=conversation.id,
