+++ backend/tests/unit/test_developer_from_segments_idempotency.py
@@ -270,16 +270,15 @@ def test_completed_desktop_meeting_persists_exact_conversation_arrival(monkeypat
 
     monkeypatch.setattr(developer, 'process_conversation', _process)
     arrival = MagicMock()
-    monkeypatch.setattr(developer, 'persist_capture_arrival_intent', arrival)
+    monkeypatch.setattr(developer, 'persist_desktop_meeting_arrival', arrival)
 
     response = developer._create_conversation_from_segments(
         'uid1', _request(client_session_id='meeting-session-1', conversation_role='meeting')
     )
 
     assert response.id == expected_id
-    arrival.assert_called_once_with(
-        'uid1', conversation_id=expected_id, summary='Design review', is_desktop_meeting=True
-    )
+    assert arrival.call_args.args[0] == 'uid1'
+    assert arrival.call_args.args[1].id == expected_id
 
 
 def test_client_session_id_retry_returns_existing_without_processing(monkeypatch):
@@ -319,7 +318,7 @@ def test_completed_desktop_meeting_retry_repairs_missing_arrival(monkeypatch):
     process = MagicMock()
     monkeypatch.setattr(developer, 'process_conversation', process)
     arrival = MagicMock()
-    monkeypatch.setattr(developer, 'persist_capture_arrival_intent', arrival)
+    monkeypatch.setattr(developer, 'persist_desktop_meeting_arrival', arrival)
 
     response = developer._create_conversation_from_segments(
         'uid1', _request(client_session_id='meeting-session-1', conversation_role='meeting')
@@ -327,9 +326,7 @@ def test_completed_desktop_meeting_retry_repairs_missing_arrival(monkeypatch):
 
     assert response.id == expected_id
     process.assert_not_called()
-    arrival.assert_called_once_with(
-        'uid1', conversation_id=expected_id, summary='Design review', is_desktop_meeting=True
-    )
+    arrival.assert_called_once_with('uid1', conversations_db.get_conversation.return_value)
 
 
 def test_completed_ambient_retry_cannot_reclassify_conversation_as_meeting(monkeypatch):
@@ -350,13 +347,13 @@ def test_completed_ambient_retry_cannot_reclassify_conversation_as_meeting(monke
     )
     monkeypatch.setattr(developer, 'process_conversation', MagicMock())
     arrival = MagicMock()
-    monkeypatch.setattr(developer, 'persist_capture_arrival_intent', arrival)
+    monkeypatch.setattr(developer, 'persist_desktop_meeting_arrival', arrival)
 
     developer._create_conversation_from_segments(
         'uid1', _request(client_session_id='ambient-session-1', conversation_role='meeting')
     )
 
-    arrival.assert_not_called()
+    arrival.assert_called_once_with('uid1', conversations_db.get_conversation.return_value)
 
 
 def test_client_session_id_concurrent_claim_loser_returns_existing_without_processing(monkeypatch):
